Another attempt.
Now, accidentally the game that was uploaded quite a while ago shows up in the preorder page because the month is set to "01" instead of "1". But that is not the key point. The one I have changed does not show up. Now, that product will be "released" in Feb. Lets see how it works.