Replace old mailchimp subscribe box to mailerlite one.
It is time to get rid of old non-functional subscribe box and replace it with a new one. Preliminary testing shows HTML version of the code works. But padding needs to be adjusted. Since design itself does not allow extra padding, let me see if applying additional formatting would improve the situation.