A few more poking around sm-simple.css.
After countless poking around finding, maybe I can start trying to style the top list so it will not display that white bezel once and for all. But I honestly don't think that could be achieved unless I change how the menu is displayed, modifying a ton of other codes. We will see.