Planning to rewrite top bar CSS
Tried to change many possible colors but none of them affects that ugly white rectangle. At this point, I dont really want to change other colors outside of that single CSS. What if that white background is for website background? Actually, as proof of concept, might just give it a shot.